requirements.txt
asn1crypto==1.5.1
backports.tarfile==1.2.0
boto3==1.38.13
botocore==1.38.13
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==44.0.3
dotenv==0.9.9
exceptiongroup==1.3.0
filelock==3.18.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jmespath==1.0.1
keyring==25.6.0
more-itertools==10.7.0
packaging==25.0
platformdirs==4.3.8
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0
